Convert AvailableArticlesList and AvailableArticles to functional stateless components
