+++ app/assets/javascripts/components/articles/available_articles.jsx
@@ -5,7 +5,7 @@ import _ from 'lodash';
 
 import AssignCell from '../students/assign_cell.jsx';
 import AvailableArticle from './available_article.jsx';
-import AvailableArticlesList from '../articles/available_article_list.jsx';
+import AvailableArticlesList from '../articles/available_articles_list.jsx';
 import AssignmentStore from '../../stores/assignment_store.js';
 
 function getState() {
